% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adapt_rwmh_metrop.R
\name{adapt_rwmh_metrop}
\alias{adapt_rwmh_metrop}
\title{Simulate adaptive RWMH algorithm (SCAM, Atchade and RAMA method)}
\usage{
adapt_rwmh_metrop(target_pdf, base, nsamples, burn_in = NULL, a_rama = NULL,
  b_rama = NULL, M_rama = NULL, def.scale = 1, verb = TRUE,
  atchade_low = NULL, atchade_high = NULL, method = c("Atchade", "SCAM",
  "Rama"))
}
\arguments{
\item{target_pdf}{The target density function from which the user wants to generate samples.}

\item{base}{The starting value of the chain}

\item{nsamples}{The number of samples to be drawn using the RWMH algorithm.}

\item{burn_in}{The number of samples assigned as burn-in period. The default burn-in is taken to be one-third of nsamples.}

\item{a_rama}{The scaling used in RAMA if norm of the chain at current iterate is less than dimension.}

\item{b_rama}{The scaling used in RAMA if norm of the chain at current iterate is greater than dimension.}

\item{M_rama}{The bounds on a_rama and b_rama, lower bound is -M_rama, upper bound is M_rama.}

\item{atchade_low}{The lower bound of scale for Atchade scheme}

\item{atchade_high}{The upper bound of scale for Atchade scheme}

\item{method}{The method of adaptation used for generating the chain. May be one of 3 types- SCAM,
              Atchade and the RAMA methods.}

\item{scale}{The proposal density scaling parameter. An approximation of the optimal scaling given the target_pdf is performed by OptimalScaling().
The default scale is this estimated optimal scaling}
}
\description{
The function simulates adaptive RWMH chain of length nsamples using one of the three methods of adaptation - Haario, Atchade and RAMA.
             The other inputs and the output same as rwmh_metrop function.
}
\author{
Kushal K Dey
}

