% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclapply.hack.R
\name{mclapply.hack}
\alias{mclapply.hack}
\title{A hacking function for mclapply in Windows}
\usage{
mclapply.hack(...)
}
\description{
This function mimics the mclapply function in Windows machine using makeCluster and parapply functionalities.
}
\author{
Nathan VanHoudnos
}

