% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mclapply_hack.R
\name{mclapply_hack}
\alias{mclapply_hack}
\title{A hacking function for mclapply in Windows}
\usage{
mclapply_hack(...)
}
\description{
This function mimics the mclapply function in Windows machine using makeCluster and parapply functionalities.
}
\author{
Nathan VanHoudnos, Kushal K Dey
}

