% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multry_rwmh_metrop.R
\name{mt_rwmh_metrop}
\alias{mt_rwmh_metrop}
\title{Simulate a Multiple Try RandomWalk Metropolis Hastings (MT-RWMH) algorithm (Liu, Liang and Wong 2000)}
\usage{
mt_rwmh_metrop(target_pdf, scale, base, nmove, nsamples, burn_in = NULL,
  verb = FALSE)
}
\arguments{
\item{target_pdf}{The log target density function from which the user wants to generate samples.}

\item{scale}{The proposal density scaling parameter. An approximation of the optimal scaling given the target_pdf is performed by OptimalScaling().
The default scale is this estimated optimal scaling}

\item{base}{The starting value of the chain}

\item{nmove}{The total number of moves proposed.}

\item{nsamples}{The number of samples to be drawn using the TMCMC algorithm.}

\item{burn_in}{The number of samples assigned as burn-in period. The default burn-in is taken to be one-third of nsamples.}

\item{verb}{logical parameter, if TRUE the function prints the progress of simulation.}
}
\value{
Returns a list containing the following items
\item{chain}{The full chain produced by the MT-TMCMC method.}
\item{post.mean}{The estimated posterior mean of the TMCMC chain adjusting for burn-in.}
 @author  Kushal K Dey

 @useDynLib tmcmcR
 @import tmcmcR
 @export
}
\description{
The function simulates a Multiple try RWMH (MT-RWMH) chain of length nsamples using the scale, base and
burn in taken optimally as default or specified by user. The function outputs the full chain as well as the estimated
posterior mean estimated from the samples drawn post burn-in.
}

