% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rmc3.R
\name{rmc3}
\alias{rmc3}
\title{Simulate a MC3/RMC3 algorithm (Rcpp sped up version)}
\usage{
rmc3(target_pdf, beta_set, scale, base, nsamples, cycle, verb = TRUE,
  swap_adjacent = TRUE, burn_in = NULL)
}
\arguments{
\item{target_pdf}{The log target density function from which the user wants to generate samples.}

\item{scale}{The proposal density scaling parameter. An approximation of the optimal scaling given the target_pdf is performed by OptimalScaling().
The default scale is this estimated optimal scaling}

\item{base}{The starting value of the chain}

\item{nsamples}{The number of samples to be drawn using the TMCMC algorithm.}

\item{burn_in}{The number of samples assigned as burn-in period. The default burn-in is taken to be one-third of nsamples.}
}
\description{
The function simulates a MC3/RMC3 chain of length nsamples using the scale, base and burn in taken optimally as default or specified by user.
 beta_set is the set of inverse temperatures chosen using select_inverse_temp() function,
 either under fixed scheme (MC3) or under randomized scheme (RMC3)
}
\author{
Kushal K Dey
}

