% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rwmh_metrop.R
\name{rwmh_metrop}
\alias{rwmh_metrop}
\title{Simulate a Random Walk Metropolis Hastings algorithm (Rcpp sped up version)}
\usage{
rwmh_metrop(target_pdf, scale, base, nsamples, burn_in = NULL, verb = TRUE)
}
\arguments{
\item{target_pdf}{The target density function from which the user wants to generate samples.}

\item{scale}{The proposal density scaling parameter. An approximation of the optimal scaling given the target_pdf is performed by OptimalScaling().
The default scale is this estimated optimal scaling}

\item{base}{The starting value of the chain}

\item{nsamples}{The number of samples to be drawn using the RWMH algorithm.}

\item{burn_in}{The number of samples assigned as burn-in period. The default burn-in is taken to be one-third of nsamples.}

\item{verb}{logical parameter, if TRUE the function prints the progress of simulation.}
}
\value{
Returns a list containing the following items
\item{chain}{The full chain produced by the RWMH method.}
\item{post.mean}{The estimated posterior mean of the RWMH chain adjusting for burn-in.}
}
\description{
The function simulates a RWMH chain of length nsamples using the scale, base and burn in taken optimally as default or specified by user. The function
outputs the full chain as well as the estimated posterior mean estimated from the samples drawn post burn-in.
}
\author{
Kushal K Dey
}

