% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/select_inverse_temp.R
\name{select_inverse_temp}
\alias{select_inverse_temp}
\title{Select inverse temperatures in RMC3 and RTMC3}
\usage{
select_inverse_temp(pdf_component, minbeta = 0.05, L_iter = 50,
  sim_method = c("RWMH", "TMCMC"), inv_temp_scheme = c("randomized",
  "fixed"))
}
\arguments{
\item{pdf_component}{The marginal distribution of the iid product component target distribution}

\item{minbeta}{the minimum cut off of inverse temperature we allow in RMC3 or RTMC3 (default is 0.05)}

\item{L_iter}{The number of sub-iterations required to fix each iterate of the inverse temperatures}

\item{method}{The method used to simulate from the marginal pdf component. Choices include TMCMC and RWMH.}
}
\description{
The function selects the inverse temperatures using a Stochastic approximation
algorithm for RMC3 or RTMC3 chains.
}
\author{
Kushal K Dey
}

