% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tmcmc_metrop.R
\name{tmcmc_metrop}
\alias{tmcmc_metrop}
\title{Simulate a Transformation based Markov Chain Monte Carlo (TMCMC) algorithm (Rcpp sped up version)}
\usage{
tmcmc_metrop(target_pdf, scale, base, nsamples, burn_in = NULL)
}
\arguments{
\item{target_pdf}{The target density function from which the user wants to generate samples.}

\item{scale}{The proposal density scaling parameter. An approximation of the optimal scaling given the target_pdf is performed by OptimalScaling().
The default scale is this estimated optimal scaling}

\item{base}{The starting value of the chain}

\item{nsamples}{The number of samples to be drawn using the TMCMC algorithm.}

\item{burn_in}{The number of samples assigned as burn-in period. The default burn-in is taken to be one-third of nsamples.}
}
\description{
The function simulates a TMCMC chain of length nsamples using the scale, base and burn in taken optimally as default or specified by user. The function
outputs the full chain as well as the estimated posterior mean estimated from the samples drawn post burn-in.
}
\author{
Kushal K Dey, Sourabh Bhattacharya
}

