% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adapt_tmcmc_metrop.R
\name{adapt_tmcmc_metrop}
\alias{adapt_tmcmc_metrop}
\title{Simulate adaptive TMCMC algorithm (SCAM, Atchade and RAMA method)}
\usage{
adapt_tmcmc_metrop(target_pdf, base, nsamples, def.scale = 1,
  burn_in = NULL, a_rama = NULL, b_rama = NULL, M_rama = NULL,
  atchade_low = NULL, atchade_high = NULL, method = c("Atchade", "SCAM",
  "Rama"), verb = TRUE)
}
\arguments{
\item{target_pdf}{The target density function from which the user wants to generate samples.}

\item{base}{The starting value of the chain}

\item{nsamples}{The number of samples to be drawn using the TMCMC algorithm.}

\item{def.scale}{the initial scale of the proposal chosen, on which we update adaptively. Default is 1.}

\item{burn_in}{The number of samples assigned as burn-in period. The default burn-in is taken to be one-third of nsamples.}

\item{a_rama}{The scaling used in RAMA if norm of the chain at current iterate is less than dimension.}

\item{b_rama}{The scaling used in RAMA if norm of the chain at current iterate is greater than dimension.}

\item{M_rama}{The bounds on a_rama and b_rama, lower bound is -M_rama, upper bound is M_rama.}

\item{atchade_low}{The lower bound of scale for Atchade scheme}

\item{atchade_high}{The upper bound of scale for Atchade scheme}

\item{method}{The method of adaptation used for generating the chain. May be one of 3 types- SCAM,
Atchade and the RAMA methods.}

\item{verb}{logical parameter, if TRUE the function prints the progress of simulation.}
}
\value{
Returns a list containing the following items
\item{chain}{The full chain produced by the adaptive TMCMC method with user-chosen method of adaptation.}
\item{post.mean}{The estimated posterior mean of the adaptive TMCMC chain adjusting for burn-in.}
}
\description{
The function simulates adaptive TMCMC chain of length nsamples using one of the three
             methods of adaptation - SCAM, Atchade and RAMA (analogously defined as in RWMH case).
}
\author{
Kushal K Dey
}

