% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/select_inverse_temp.R
\name{select_inverse_temp}
\alias{select_inverse_temp}
\title{Select inverse temperatures in RMC3 and RTMC3}
\usage{
select_inverse_temp(pdf_component, minbeta = 0.01, L_iter = 50,
  sim_method = c("RWMH", "TMCMC"), inv_temp_scheme = c("randomized",
  "fixed"), rho_start = 0, scale = 0.1)
}
\arguments{
\item{pdf_component}{The univariate marginal distribution of the iid product component target distribution}

\item{minbeta}{the cut off of inverse temperature which when crossed, we stop at that value of inverse
temperature. Helps to check the lower bound of the inverse temperatures. Default is 0.01.}

\item{L_iter}{The number of sub-iterations required to fix each iterate of the inverse temperatures}

\item{sim_method}{The method used to simulate from the marginal pdf component. Choices include TMCMC and RWMH.}

\item{inv_temp_scheme}{The selection of inverse temperatures scheme followed, affects the stochastic approximation
driving the inverse temperature updates. Choices include "fixed" and "randomized".}

\item{rho_start}{The scale update for the inverse temperature selection. Default is 0.}

\item{scale}{The stochastic approximation scaling parameter. Default is 0.1. The smaller this value,
       more inverse temperatures will be selected and algorithm will be more fine-grained.}
}
\value{
Returns a vector of inverse temperatures starting from 1 and ending at the first value of inverse
       temperature smaller than minbeta.
}
\description{
The function selects the inverse temperatures using a Stochastic approximation
algorithm for RMC3 or RTMC3 chains.
}
\author{
Kushal K Dey
}

